[chore] moved the logs endpoint to the run module and registered it directly to the app
